Shiny UI for data frame viewer application

This is a simple Shiny application to demonstrate the [runAppWithParams()] function. This will display the structure and contents of any data frames listed in the `data_frames` list object.

Usage

df_viewer_ui

df_viewer_server(input, output, session)

Format

An object of class shiny.tag.list (inherits from list) of length 4.

Arguments

input

input object from Shiny.

output

output object from Shiny.

session

session object from Shiny.

Examples

if (interactive()) { # Only run this example in interactive R sessions
data(mtcars)
data(faithful)
runAppWithParams(ui = ShinyDemo::df_viewer_ui,
                 server = ShinyDemo::df_viewer_server,
                 data_frames = list(mtcars = mtcars, faithful = faithful),
                 port = 2112)
}